requirements.txt
PyQt6>=6.6.0
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
numpy>=1.26.0
PyYAML>=6.0.1
Pillow>=10.0.0
